refactor: move apply-dark-mode hook into other directory
