+++ frontend/src/hooks/dark-mode/use-apply-dark-mode.ts
@@ -1,12 +1,12 @@
 /*
- * SPDX-FileCopyrightText: 2022 The HedgeDoc developers (see AUTHORS file)
+ * SPDX-FileCopyrightText: 2023 The HedgeDoc developers (see AUTHORS file)
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { DarkModePreference } from '../../redux/dark-mode/types'
 import { isClientSideRendering } from '../../utils/is-client-side-rendering'
 import { Logger } from '../../utils/logger'
-import { useApplicationState } from './use-application-state'
+import { useApplicationState } from '../common/use-application-state'
 import useMediaQuery from '@restart/hooks/useMediaQuery'
 import { useEffect } from 'react'
