+++ frontend/src/pages/cheatsheet.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { CheatsheetContent } from '../components/editor-page/app-bar/cheatsheet/cheatsheet-content'
-import { useApplyDarkMode } from '../hooks/common/use-apply-dark-mode'
+import { useApplyDarkMode } from '../hooks/dark-mode/use-apply-dark-mode'
 import type { NextPage } from 'next'
 import { Container } from 'react-bootstrap'
